requirements.txt
opencv-python
mediapipe
pyautogui
pyaudio
websockets
python-dotenv
pystray
Pillow
customtkinter
numpy
